+++ examples/lexicons/codelingo/ast/typescript/src/babel/ParserBabel.ts
@@ -3,9 +3,10 @@ import { File, Node } from "@babel/types";
 import path from "path";
 import { KeyManager } from "../common/KeyManager";
 import { AstNode, Dictionary, EmitterFn, NAMESPACE } from "../common/model";
-import { EmitInstructions, shapeNodeForEmit } from "./NodeShaper";
 import { makeProperties } from "../common/property";
 import { makeCommonPropertiesBabel } from "./helpers";
+import { EmitInstructions } from "./model";
+import { shapeNodeForEmit } from "./NodeShaper";
 
 const COMMON_PLUGINS: babelParser.ParserPlugin[] = [
     "decorators-legacy", // XXX: "decorators" plugin requires a "decoratorsBeforeExport" option -- which we don't know in advance
